+++ tests/testthat/test_components.R
@@ -27,22 +27,17 @@ for (data_type in data_types) {
   qc_metrics_tables <-
     get_qc_metrics(pg_data, sample_qc_metrics, sample_sheet)
 
-  es_data <- structure(
-    list(
-      params = list(
-        control_markers = c("mIgG1", "mIgG2a", "mIgG2b")
-      ),
-      samplesheet = sample_sheet,
-      sample_aliases =
-        pixelatorES:::.sample_aliases_from_samplesheet(sample_sheet),
-      effective_samplesheet = sample_sheet,
-      file_paths = file_paths,
-      pxl_data_processed = pg_data,
-      qc_raw = sample_qc_metrics,
-      qc = qc_metrics_tables,
-      proximity = NULL
+  es_data <- test_es_data(
+    params = list(
+      control_markers = c("mIgG1", "mIgG2a", "mIgG2b")
     ),
-    class = c("es_data", "list")
+    samplesheet = sample_sheet,
+    effective_samplesheet = sample_sheet,
+    file_paths = file_paths,
+    pxl_data_processed = pg_data,
+    qc_raw = sample_qc_metrics,
+    qc = qc_metrics_tables,
+    proximity = NULL
   )
 
   test_message <- paste("Components work as expected for", data_type, "data")